% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madrid.air.R
\docType{package}
\name{madrid.air-package}
\alias{madrid.air}
\alias{madrid.air-package}
\title{Madrid Air Quality Data Parser}
\description{
Convert raw air quality data published by
\url{http://datos.madrid.es/} to a clean data frame.
}
\section{Functions}{

\code{\link{madrid.air.parse}}
}

\section{Datasets}{

\code{\link{madrid.air.magnitudes}}
\code{\link{madrid.air.stations}}
}
\author{
Ramon Novoa \email{ramonnovoa@gmail.com}
}
\seealso{
\url{https://github.com/nramon/madrid.air}
}

