% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madrid.air.magnitudes.R
\docType{data}
\name{madrid.air.magnitudes}
\alias{madrid.air.magnitudes}
\title{Monitored physical magnitudes}
\format{A data frame with 27 observations of 4 variables:
\describe{
 \item{Name}{Magnitude name (e.g., "Carbon monoxide").}
 \item{Abbreviation}{Magnitude abbreviation (e.g., "PM10").}
 \item{Magnitude}{Numeric code used to refer to this magnitude in the raw
data.}
 \item{Unit}{Measurement unit (e.g., "m/s").}
}}
\source{
\url{http://datos.madrid.es/FWProjects/egob/contenidos/datasets/ficheros/MedioAmbiente_CalidadAire/INTPHORA-DIA_V2.2.pdf}
}
\usage{
data(madrid.air.magnitudes)
}
\description{
A dataset containing information about the physical magnitudes monitored by
the air quality monitoring stations located in the city of Madrid.
}
\keyword{datasets}

