% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madrid.air.parse.R
\name{madrid.air.parse}
\alias{madrid.air.parse}
\title{Air quality data parser.}
\usage{
madrid.air.parse(input, output = NA, station = NA, magnitude = NA)
}
\arguments{
\item{input}{Input file containing the raw air quality data.}

\item{output}{Optional CSV file where the air quality data data frame will
be saved.}

\item{station}{Filter data by station code (see
\code{\link{madrid.air.stations}}).}

\item{magnitude}{Filter data by magnitude name (see
\code{\link{madrid.air.magnitudes}}).}
}
\value{
A data frame containing air quality data.
}
\description{
Parses raw air quality data published by \url{http://datos.madrid.es/} and returns
a clean data frame.
}
\examples{
# Download and parse daily data.
download.file('http://datos.madrid.es/egob/catalogo/201410-14-calidad-aire-diario.txt',
              '201410-14-calidad-aire-diario.txt')
madrid.air.parse('201410-14-calidad-aire-diario.txt')

# Save air quality data to disk.
madrid.air.parse('201410-14-calidad-aire-diario.txt',
                 output='201410-14-calidad-aire-diario.csv')

# Download and parse hourly data.
download.file('http://datos.madrid.es/egob/catalogo/201200-29-calidad-aire-horario.zip',
              '201200-29-calidad-aire-horario.zip')
unzip('201200-29-calidad-aire-horario.zip')
madrid.air.parse('Ene_mo01.txt')

# Check station information.
data(madrid.air.stations)
madrid.air.stations

# Filter by station.
madrid.air.parse('Ene_mo01.txt', station=28079001)

# Check magnitude information.
data(madrid.air.magnitudes)
madrid.air.magnitudes

# Filter by magnitude.
madrid.air.parse('Ene_mo01.txt', magnitude='CO')
}

