% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madrid.air.stations.R
\docType{data}
\name{madrid.air.stations}
\alias{madrid.air.stations}
\title{Air quality monitoring stations.}
\format{A data frame with 35 observations of 5 variables:
\describe{
 \item{Station}{Station code (e.g., "28079001").}
 \item{Location}{Address where the station is physically located.}
 \item{Longitude}{Location longitude in degrees.}
 \item{Latitude}{Location latitude in degrees.}
 \item{Comments}{Additional comments (e.g., "Closed on 04/05/2009").}
}}
\source{
\url{http://datos.madrid.es/FWProjects/egob/contenidos/datasets/ficheros/MedioAmbiente_CalidadAire/INTPHORA-DIA_V2.2.pdf}

\url{http://www.eea.europa.eu/data-and-maps/data/airbase-the-european-air-quality-database-8}
}
\usage{
data(madrid.air.stations)
}
\description{
A dataset containing information about the air quality monitoring stations
located in the city of Madrid.
}
\keyword{datasets}

